requirements.txt
Babel==2.5.3
cssselect2==0.2.1
defusedxml==0.5.0
Django==2.0
django-countries==5.0
django-easy-pdf==0.1.1
django-language-field==0.0.3
django-numpy==1.0.1
django-phonenumber-field==1.3.0
django-widget-tweaks==1.4.1
django-wkhtmltopdf==3.1.0
html5lib==1.0b10
httplib2==0.10.3
olefile==0.44
pdfrw==0.4
phonenumberslite==8.8.8
Pillow==4.3.0
pkg-resources==0.0.0
psycopg2==2.7.3.2
pycparser==2.18
PyPDF2==1.26.0
Pyphen==0.9.4
python-dateutil==2.6.1
pytz==2017.3
reportlab==3.4.0
six==1.11.0
tinycss2==0.6.1
webencodings==0.5.1
xhtml2pdf==0.2b1
